pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
version = "0.1.0"
name = "voice2text-deltabot"
description = "Delta Chat bot to extract text from voice messages"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE.txt"}
keywords = ["deltachat", "bot"]
authors = [
  {name = "adbenitez", email = "adbenitez@hispanilandia.net"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]
dependencies = [
    "deltabot-cli",
    "deltachat-rpc-server",
    "openai-whisper"
]

[project.optional-dependencies]
dev = [
  "black",
  "mypy",
  "isort",
  "pylint",
  "pylama",
  "pytest",
]

[project.scripts]
voice2text-bot = "voice2text_deltabot:main"

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = "True"
